+++ temba/flows/models.py
@@ -2,7 +2,7 @@ import logging
 import time
 from array import array
 from collections import defaultdict
-from datetime import datetime, timedelta
+from datetime import datetime
 
 import iso8601
 import pytz
@@ -487,12 +487,16 @@ class Flow(TembaModel):
             except FlowException:  # pragma: no cover
                 pass
 
-    def get_icon(self):
-        if self.flow_type == Flow.TYPE_MESSAGE:
-            return "message-square"
-        elif self.flow_type == Flow.TYPE_VOICE:
-            return "phone"
-        return "flow"
+    def get_attrs(self):
+        icon = (
+            "message-square"
+            if self.flow_type == Flow.TYPE_MESSAGE
+            else "phone"
+            if self.flow_type == Flow.TYPE_VOICE
+            else "flow"
+        )
+
+        return {"icon": icon, "type": self.flow_type}
 
     def get_category_counts(self):
         keys = [r["key"] for r in self.metadata["results"]]
@@ -1089,7 +1093,7 @@ class FlowSession(models.Model):
     uuid = models.UUIDField(unique=True)
     org = models.ForeignKey(Org, related_name="sessions", on_delete=models.PROTECT)
     contact = models.ForeignKey("contacts.Contact", on_delete=models.PROTECT, related_name="sessions")
-    status = models.CharField(max_length=1, choices=STATUS_CHOICES, null=True)
+    status = models.CharField(max_length=1, choices=STATUS_CHOICES)
 
     # the modality of this session
     session_type = models.CharField(max_length=1, choices=Flow.TYPE_CHOICES, default=Flow.TYPE_MESSAGE)
@@ -1114,6 +1118,7 @@ class FlowSession(models.Model):
     wait_started_on = models.DateTimeField(null=True)  # when it started waiting
     timeout_on = models.DateTimeField(null=True)  # when it should timeout (set by courier when last msg is sent)
     wait_expires_on = models.DateTimeField(null=True)  # when waiting run can be expired
+    wait_resume_on_expire = models.BooleanField()  # whether wait expiration can resume a parent run
 
     # the flow of the waiting run
     current_flow = models.ForeignKey("flows.Flow", related_name="sessions", null=True, on_delete=models.PROTECT)
@@ -1132,6 +1137,9 @@ class FlowSession(models.Model):
             return self.output
 
     def release(self):
+        for run in self.runs.all():
+            run.release()
+
         self.delete()
 
     def __str__(self):  # pragma: no cover
@@ -1148,6 +1156,13 @@ class FlowSession(models.Model):
                 name="flows_session_voice_expires",
                 fields=("wait_expires_on",),
                 condition=Q(session_type=Flow.TYPE_VOICE, status="W", wait_expires_on__isnull=False),
+            ),
+        ]
+        constraints = [
+            # ensure that waiting sessions have a wait started and expires
+            models.CheckConstraint(
+                check=~Q(status="W") | Q(wait_started_on__isnull=False, wait_expires_on__isnull=False),
+                name="flows_session_waiting_has_started_and_expires",
             ),
         ]
 
@@ -1215,11 +1230,6 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
     # session this run belongs to (can be null if session has been trimmed)
     session = models.ForeignKey(FlowSession, on_delete=models.PROTECT, related_name="runs", null=True)
 
-    # for an IVR session this is the connection to the IVR channel
-    connection = models.ForeignKey(
-        "channels.ChannelConnection", on_delete=models.PROTECT, related_name="runs", null=True
-    )
-
     # when this run was created, last modified and exited
     created_on = models.DateTimeField(default=timezone.now)
     modified_on = models.DateTimeField(default=timezone.now)
@@ -1233,9 +1243,6 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
 
     # if this run is part of a Surveyor session, the user that submitted it
     submitted_by = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.PROTECT, null=True, db_index=False)
-
-    # UUID of the parent run (if any)
-    parent_uuid = models.UUIDField(null=True)
 
     # results collected in this run keyed by snakified result name
     results = JSONAsTextField(null=True, default=dict)
@@ -1253,12 +1260,6 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
     is_active = models.BooleanField(default=True)
     exit_type = models.CharField(null=True, max_length=1, choices=EXIT_TYPE_CHOICES)
 
-    # TODO to be replaced by FlowSession.wait_expires_on
-    expires_on = models.DateTimeField(null=True)
-
-    # TODO drop once mailroom is no longer writing
-    events = JSONField(null=True)
-
     def release(self, delete_reason=None):
         """
         Permanently deletes this flow run
@@ -1268,9 +1269,6 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
                 self.delete_reason = delete_reason
                 self.save(update_fields=["delete_reason"])
 
-            # and any recent runs
-            self.recent_runs.all().delete()
-
             if (
                 delete_reason == FlowRun.DELETE_FOR_USER
                 and self.session is not None
@@ -1279,17 +1277,6 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
                 mailroom.queue_interrupt(self.org, session=self.session)
 
             self.delete()
-
-    def update_expiration(self, point_in_time):
-        """
-        Set our expiration according to the flow settings
-        """
-        if self.flow.expires_after_minutes:
-            self.expires_on = point_in_time + timedelta(minutes=self.flow.expires_after_minutes)
-            self.modified_on = timezone.now()
-
-            # save our updated fields
-            self.save(update_fields=["expires_on", "modified_on"])
 
     def as_archive_json(self):
         def convert_step(step):
@@ -1578,23 +1565,6 @@ class FlowPathCount(SquashableModel):
 
     class Meta:
         index_together = ["flow", "from_uuid", "to_uuid", "period"]
-
-
-class FlowPathRecentRun(models.Model):
-    """
-    TODO drop once we stop writing these
-    """
-
-    id = models.BigAutoField(primary_key=True)
-    from_uuid = models.UUIDField()
-    from_step_uuid = models.UUIDField()
-    to_uuid = models.UUIDField()
-    to_step_uuid = models.UUIDField()
-    run = models.ForeignKey(FlowRun, on_delete=models.PROTECT, related_name="recent_runs")
-    visited_on = models.DateTimeField(default=timezone.now)
-
-    class Meta:
-        indexes = [models.Index(fields=["from_uuid", "to_uuid", "-visited_on"])]
 
 
 class FlowNodeCount(SquashableModel):